replaced highlight with promo block, better naming there, added popup and close feature to promo blocks popup variant.
