--- blocks/highlight/highlight.css
@@ -1,136 +1,0 @@
-.highlight {
-    position: relative;
-    border: var(--border-m) solid var(--blue);
-    border-radius: 24px;
-    padding: var(--horizontal-spacing);
-    margin-bottom: calc(var(--spacing-xxxl) + var(--spacing-xxl));
-}
-
-.highlight > div {
-    display: flex;
-    flex-flow: column;
-}
-
-.highlight > div:nth-child(1) img {
-    width: 100%;
-    height: 250px;
-    object-fit: contain;
-}
-
-.highlight > div .icon svg {
-    width: 100%;
-    height: 250px;
-}
-
-.highlight > div:nth-child(2) {
-    position: absolute;
-    top: -60px;
-    left: -40px;
-    width: 140px; 
-    height: 140px;
-}
-
-.highlight > div:nth-child(2) > div {
-    position: relative;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.highlight > div:nth-child(1) > div:nth-child(1) {
-    display: flex;
-    justify-content: center;
-    min-width: 250px;
-}
-
-.highlight > div:nth-child(1) > div:nth-child(2) {
-    padding-top: var(--horizontal-spacing);
-}
-
-.highlight > div:nth-child(1) > div > p.button-wrapper {
-   display: flex;
-   justify-content: center;
-}
-
-.highlight > div:nth-child(1) > div > p.button-wrapper a {
-    font-size: var(--heading-size-s);
-    font-family: var(--heading-font-family);
-    padding: var(--spacing-s) calc(var(--spacing-xxxl) + var(--spacing-l));
-    margin-top: var(--horizontal-spacing);
-}
-
-.highlight .close-button {
-    font-size: var(--heading-size-xxl);
-    font-family: var(--heading-font-family);
-    color: var(--blue);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: absolute;
-    top: 0;
-    right: 0;
-    margin: var(--spacing-m);
-    height: 60px;
-    width: 60px;
-    line-height: var(--line-hight-s);
-    padding: 10px;
-}
-
-.highlight .close-button:hover {
-    background-color: var(--blue-lightest);
-    border-radius: 50%;
-}
-
-.highlight > div:nth-child(2) > div > p {
-    position: absolute;
-    color: var(--blue);
-    font-weight: 600;
-    z-index: 3;
-    max-width: 50px;
-    text-align: center;
-}
-
-.highlight > div:nth-child(2) > div > div svg {
-    height: 100%;
-    width: 100%;
-    color: var(--pink);
-    animation-name: spin;
-    animation-duration: 15000ms;
-    animation-iteration-count: infinite;
-    animation-timing-function: linear; 
-}
-
-@keyframes spin {
-    from {
-        transform:rotate(0deg);
-    }
-    
-    to {
-        transform:rotate(360deg);
-    }
-}
-
-@media (width >= 900px) {
-    .highlight > div {
-        flex-flow: row;
-    } 
-
-    .highlight > div:nth-child(1) > div:nth-child(1) {
-        padding-right: var(--horizontal-spacing);
-    }
-
-    .highlight > div:nth-child(1) > div:nth-child(2) {
-        width: 200px;
-        padding-right: var(--horizontal-spacing);
-        flex-grow: 1;
-        padding-top: 0;
-    }
-
-    .highlight > div:nth-child(1) > div:nth-child(2) h3 {
-        font-size: var(--heading-size-xl);
-    }
-
-    .highlight > div:nth-child(1) > div > p.button-wrapper {
-        justify-content: flex-start;
-    }
-}
