+++ blocks/promo/promo.css
@@ -1,0 +1,144 @@
+
+.promo.popup {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100vw;
+    height: 100vh;
+    background-color: #0B21E330;
+    z-index: 5;
+}
+
+.promo > .promo-container {
+    position: relative;
+    border: var(--border-m) solid var(--blue);
+    border-radius: 24px;
+    padding: var(--horizontal-spacing);
+}
+
+.promo.popup > .promo-container {
+    position: fixed;
+    z-index: 6;
+    top: 50%;
+    left: 50%;
+    transform:translate(-50%,-50%);
+    min-height: fit-content;
+    min-width: 70vw;
+    max-width: 800px;
+    overflow: auto;
+    background-color: var(--white);
+    border: 2px solid var(--blue);
+    border-radius: 12px;
+}
+
+.promo > .promo-container > div {
+    display: flex;
+    flex-flow: column;
+}
+
+.promo > .promo-container > div:nth-child(1) img {
+    width: 100%;
+    height: 250px;
+    object-fit: contain;
+    margin-top: 60px;
+}
+
+.promo > .promo-container > div .icon svg {
+    width: 100%;
+    height: 250px;
+}
+
+.promo > .promo-container > div:nth-child(2) {
+    position: absolute;
+    top: -60px;
+    left: -40px;
+    width: 140px; 
+    height: 140px;
+}
+
+.promo > .promo-container > div:nth-child(2) > div {
+    position: relative;
+    z-index: 10;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.promo > .promo-container > div:nth-child(1) > div:nth-child(1) {
+    display: flex;
+    justify-content: center;
+    min-width: 250px;
+}
+
+.promo > .promo-container > div:nth-child(1) > div:nth-child(2) {
+    padding-top: var(--horizontal-spacing);
+}
+
+.promo > .promo-container > div:nth-child(1) > div > p.button-wrapper {
+    display: flex;
+    justify-content: center;
+}
+
+.promo > .promo-container > div:nth-child(1) > div > p.button-wrapper a {
+    font-size: var(--heading-size-s);
+    font-family: var(--heading-font-family);
+    padding: var(--spacing-s) calc(var(--spacing-xxxl) + var(--spacing-l));
+    margin-top: var(--horizontal-spacing);
+}
+
+.promo > .promo-container .close-button {
+    font-size: var(--heading-size-xxl);
+    font-family: var(--heading-font-family);
+    color: var(--blue);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position: absolute;
+    top: 0;
+    right: 0;
+    margin: var(--spacing-m);
+    height: 60px;
+    width: 60px;
+    line-height: var(--line-hight-s);
+    padding: 5px;
+}
+
+.promo > .promo-container .close-button:hover {
+    background-color: var(--blue-lightest);
+    border-radius: 50%;
+}
+
+@media (width >= 600px) {
+    .promo > .promo-container > div:nth-child(1) img {
+        margin-top: 0;
+    }
+}
+
+@media (width >= 900px) {
+    .promo > .promo-container > div:nth-child(1) img {
+        height: 400px;
+    }
+
+    .promo > .promo-container > div:nth-child(1) {
+        flex-flow: row;
+    }
+
+    .promo > .promo-container > div:nth-child(1) > div:nth-child(1) {
+        padding-right: var(--horizontal-spacing);
+    }
+
+    .promo > .promo-container > div:nth-child(1) > div:nth-child(2) {
+        width: 200px;
+        padding-right: var(--horizontal-spacing);
+        flex-grow: 1;
+        padding-top: 0;
+    }
+
+    .promo > .promo-container > div:nth-child(1) > div:nth-child(2) h3 {
+        font-size: var(--heading-size-xl);
+    }
+
+    .promo > .promo-container > div:nth-child(1) > div > p.button-wrapper {
+        justify-content: flex-start;
+    }
+}
